requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
email-validator==2.1.0
python-dotenv==1.0.0
itsdangerous==2.1.2
